# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Waves API
  version: 1.0.0
  summary: Read and write a Waves shared ledger — groups, expenses, settlements and balances.
  description: |
    The Waves API lets a program act for a Waves user: read the groups they are
    in, add and edit expenses, record payments, and read what everybody is owed.
    Every request runs as that person with their row-level security intact, so a
    token can never reach a group its owner could not reach from the app. A scope
    only narrows what the token may do; it never widens what the person can see.

    ## Authenticating

    There are two ways in, and they produce the same kind of credential.

    **A personal access token** is a developer acting on their own account. It is
    created in the Waves developer console, carries a fixed scope list, and is
    sent as `Authorization: Bearer wavs_pat_…`. Use this when the program and the
    account belong to the same person.

    **OAuth 2.1 authorization code with PKCE** is somebody else's program acting
    for a Waves user who approved it. Send the browser to `/oauth/authorize`,
    exchange the returned code at `/oauth/token`, and use the resulting
    `wavs_at_…` access token in the same `Authorization` header. `S256` is the
    only code challenge method accepted, and a refresh token is only issued when
    `offline_access` was granted. Refresh tokens rotate: each one is single-use,
    and replaying a spent one disconnects the application.

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/settlements:
    get:
      tags: [Settlements]
      operationId: listSettlements
      summary: List payments
      description: Most recently initiated first.
      security:
        - bearerAuth: []
        - oauth2: [settlements.read]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: group_id
          in: query
          required: false
          description: Only payments in this group.
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          description: >-
            Only payments in this state. Validated against the lifecycle: a
            value outside the enum is a 400, not an empty page.
          schema:
            $ref: '#/components/schemas/SettlementStatus'
      responses:
        '200':
          description: One page of settlements.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementPage'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [Settlements]
      operationId: createSettlement
      summary: Record a payment
      description: >-
        Records a claim that money moved, not the movement itself; it counts
        against the ledger once the payee confirms it. The `Idempotency-Key`
        matters more here than anywhere else in the API — a retried "I paid
        them" that recorded twice would take a real debt to a negative balance.
      security:
        - bearerAuth: []
        - oauth2: [settlements.write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSettlementRequest'
      responses:
        '201':
          description: The settlement as recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settlement'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422':
          description: >-
            `NOTHING_OWED` — there is no debt between these two people for this
            payment to settle.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/settlements/{settlementId}:
    parameters:
      - $ref: '#/components/parameters/SettlementId'
    get:
      tags: [Settlements]
      operationId: getSettlement
      summary: Read one payment
      security:
        - bearerAuth: []
        - oauth2: [settlements.read]
      responses:
        '200':
          description: The settlement.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settlement'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/settlements/{settlementId}/confirm:
    parameters:
      - $ref: '#/components/parameters/SettlementId'
    post:
      tags: [Settlements]
      operationId: confirmSettlement
      summary: Confirm that a payment reached you
      description: >-
        "Yes, that reached me." Only the payee may say it, and that is decided
        server-side — a token with `settlements.write` can confirm the payments
        addressed to its owner and nobody else's. Takes no body.
      security:
        - bearerAuth: []
        - oauth2: [settlements.write]
      responses:
        '200':
          description: The settlement, now confirmed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settlement'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: >-
            You are not the person this payment was made to, so you cannot
            confirm it.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: >-
            `INVALID_TRANSITION` — this payment is not in a state that can be
            confirmed. It has already been confirmed, or was cancelled or
            disputed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ──────────────────────────────────────────────────────────── friends ──
...
  /oauth/authorize:
    get:
      tags: [OAuth]
      operationId: authorize
      summary: Start the authorization code flow
      description: |
        Checks the *shape* of the request and redirects the browser to this
        deployment's own consent screen, carrying the parameters along. It never
        redirects to a caller-supplied address and never touches the database,
        which is why it cannot be used to probe whether a client id exists —
        whether the client and the redirect are real is decided later, under the
        signed-in user's session.

        Errors are returned as an RFC 6749 JSON body rather than bounced to
        `redirect_uri`: a validator that redirected errors to an unverified
        address would be an open redirect wearing an OAuth costume.
      security: []
      parameters:
        - name: response_type
          in: query
          required: true
          description: Must be `code`.
          schema:
            type: string
            enum: [code]
        - name: client_id
          in: query
          required: true
          description: The application's public client id.
          schema:
            type: string
            examples: ['wavs_app_5f2c1d9a4b7e42f1a0c6d3e8b91f7a24']
        - name: redirect_uri
          in: query
          required: true
          description: >-
            Where to deliver the code. Must be one of the addresses registered
            for this application — checked later, when the user approves.
          schema:
            type: string
            format: uri
        - name: scope
          in: query
          required: true
          description: >-
            Space-separated. Every entry must be a known scope; one unknown
            entry fails the whole request rather than being silently dropped.
          schema:
            type: string
            examples: ['identity.read groups.read expenses.write offline_access']
        - name: code_challenge
          in: query
          required: true
          description: The unpadded base64url SHA-256 of your verifier — 43 characters.
          schema:
            type: string
            pattern: '^[A-Za-z0-9_-]{43}$'
        - name: code_challenge_method
          in: query
          required: true
          description: '`S256` only. `plain` is a challenge whose answer is in the same request.'
          schema:
            type: string
            enum: [S256]
        - name: state
          in: query
          required: false
          description: Returned unchanged on the redirect back.
          schema:
            type: string
      responses:
        '302':
          description: >-
            Redirect to this deployment's consent screen. The user approves
            there, and the screen calls `POST /oauth/authorize` with their own
            session.
          headers:
            Location:
              description: The consent screen, with the request's parameters attached.
              schema:
                type: string
                format: uri
        '400':
          description: >-
            The request is malformed, or `scope` names something this deployment
            does not know.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthError' }
        '500':
          description: >-
            This deployment has no consent screen configured
            (`WAVES_API_WEB_URL` is unset).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthError' }
    post:
      tags: [OAuth]
      operationId: approveAuthorization
      summary: Record the user's approval and mint a code
      description: >-
        Called by the consent screen, authenticated by the *user's own Waves
        session* rather than a client credential — the decision being recorded
        is theirs. The redirect and the scopes are re-checked against the
        application's registration here, because the browser between the two
        screens is the attacker's ground and nothing it carries is believed
        twice. Returns the address to send the browser to, rather than
        redirecting, so the screen can control the navigation.


        Because it carries a session, this endpoint is on the console's CORS
        surface, not the public one: a browser can only call it from an origin
        in the deployment's `WAVES_API_ALLOWED_ORIGINS` allowlist, and a
        deployment that has not set that variable refuses browser calls
        entirely. The `GET` beside it carries no credential and stays public.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveAuthorizationRequest'
      responses:
        '200':
          description: Where to send the browser, with the code attached.
          content:
            application/json:
              schema:
                type: object
                required: [redirect_to]
                properties:
                  redirect_to:
                    type: string
                    format: uri
                    description: >-
                      The application's registered redirect with `code` and, if
                      you sent one, `state` appended.
        '400':
          description: Malformed, or the redirect or scopes are not registered for this client.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthError' }
        '401':
          description: No Waves session, or one the authorization server would not vouch for.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthError' }
        '500':
          description: Something went wrong on our side.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthError' }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: wavs_pat_… or wavs_at_…
      description: |
        A Waves API token in `Authorization: Bearer …`. Two kinds are accepted
        here and they behave identically once presented:

        - `wavs_pat_…` — a personal access token, created in the developer
          console, acting for the developer's own account.
        - `wavs_at_…` — an OAuth access token, acting for a user who approved an
          application.

        A refresh token (`wavs_rt_…`) is a credential for `/oauth/token` and
        nothing else; presenting one here is refused with `invalid_token`.

        The token string carries no scope list — the row does — so the scopes
        each operation lists apply to both kinds, and narrowing an
        application's scopes takes effect on its next request.

        A missing token is `401 unauthorized` with
        `WWW-Authenticate: Bearer realm="waves"`. A token that is malformed,
        expired, revoked, or belongs to a disabled application is `401
        invalid_token` with
        `WWW-Authenticate: Bearer realm="waves", error="invalid_token"`. A live
        token that lacks the scope the route needs is `403 insufficient_scope`.
    oauth2:
      type: oauth2
      description: >-
        OAuth 2.1 authorization code with PKCE (`S256` only). The URLs below are
        relative to this deployment's own base URL; the authoritative copies are
        in `/.well-known/oauth-authorization-server`, which derives them from
        the request. Ask for `offline_access` if you need a refresh token.
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          refreshUrl: /oauth/token
          scopes:
            identity.read: See your name, avatar, country, language, default currency and payment
...
  headers:
    XRequestId:
      description: >-
        This request's id. Present on every response, including errors, and
        repeated in the error body as `request_id`.
      schema:
        type: string
        format: uuid
...
    WWWAuthenticate:
      description: >-
        `Bearer realm="waves"` when no token was sent, and
        `Bearer realm="waves", error="invalid_token"` when one was sent and is
        not usable.
      schema:
        type: string

  # ───────────────────────────────────────────────── shared parameters ──
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: >-
        How many rows to return. Defaults to 25; anything above 100 is quietly
        capped at 100. A value below 1, or one that is not a whole number, is a
        400.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Cursor:
      name: cursor
      in: query
      required: false
      description: >-
        The `next_cursor` from the previous page. Opaque: it encodes a sort key
        and a tiebreaker, and treating it as anything other than a string you
        pass back will break. A cursor this deployment did not issue is a 400.
      schema:
        type: string
        examples:
          ['MjAyNi0wOS0wOFQxMDoxNDowMi4xMTFaIDNmYTg1ZjY0LTU3MTctNDU2Mi1iM2ZjLTJjOTYzZjY2YWZhNg']
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: >-
        A key of your own choosing, 8 to 200 characters, that makes this write
        safe to retry: the same key on the same route returns the original write
        rather than making a second one. (`POST /v1/categories` is the one
        exception — a retry there is refused with a 409 instead of returning the
        original row. Either way you never get a duplicate.) It is scoped to the
        token that sent it, so two developers who both pick `"1"` do not
        collide. Omitting it means the write is not idempotent, which is a
        choice, not an error — but a key outside 8–200 characters is a 400 with
        `details.header = "Idempotency-Key"`.
      schema:
        type: string
        minLength: 8
        maxLength: 200
        examples: ['dinner-2026-09-08-goa-01']
...
    SettlementId:
      name: settlementId
      in: path
      required: true
      description: The settlement's id.
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: >-
        `invalid_request` — the body, a query parameter or a header is not
        something this endpoint can act on. The message names the field.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: >-
        `unauthorized` when no credential was sent, or `invalid_token` when one
        was sent and is malformed, expired, revoked, of the wrong kind, or
        belongs to a disabled application.
      headers:
        WWW-Authenticate: { $ref: '#/components/headers/WWWAuthenticate' }
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: >-
        `insufficient_scope` when the token is live but was not granted the
        scope this route needs, or `forbidden` when the account itself may not
        do this — not a member of the group, not a party to the settlement, not
        the owner, or a guest whose trial has ended.
      headers:
        WWW-Authenticate: { $ref: '#/components/headers/WWWAuthenticate' }
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
        RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
        RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
        RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: >-
        `not_found`. Note that a row the caller is not allowed to see is absent
        rather than forbidden, so this also covers "it exists, but not for you".
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
        RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
        RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
        RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: >-
        `conflict` — the write collided with the state it was based on. A stale
        `if_revision` or `base_version_no`, shares that do not reconcile with
        the amount, or an id that is already taken.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
        RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
        RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
        RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    TooManyRequests:
      description: >-
        `rate_limited`. Two buckets are enforced, per token and per account, and
        this is returned when either is spent. Wait `Retry-After` seconds.
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
        RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
        RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
        RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    InternalError:
      description: >-
        `internal` when something went wrong on our side, or `misconfigured`
        when the deployment is missing configuration it needs (check `/health`).
        Neither says anything about the cause; quote `request_id`.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }

  # ──────────────────────────────────────────────────────────── schemas ──
  schemas:
...
    SettlementStatus:
      type: string
      enum: [initiated, confirmed, auto_confirmed, disputed, cancelled]
      description: >-
        Where a payment stands. `initiated` is a claim awaiting the payee;
        `confirmed` is the payee saying it arrived; `auto_confirmed` is the
        system doing so after a grace period; `disputed` is the payee saying it
        did not; `cancelled` is the payer withdrawing the claim.

    # ── errors ──
    Error:
      type: object
      description: >-
        Every refusal under `/v1` and `/developer` has this shape. It is nested
        under `error` so that it can never collide with a resource that happens
        to have a `message`, and so a client can branch on `"error" in body`
        without knowing which endpoint it called. Nothing internal — a database
        message, a stack, a hostname — ever reaches this body.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum:
                - invalid_request
                - unauthorized
                - invalid_token
                - insufficient_scope
                - forbidden
                - not_found
                - conflict
                - unprocessable
                - rate_limited
                - misconfigured
                - internal
              description: |
                A token to match on, rather than a sentence to parse. The HTTP
                status each one carries:

                | code | status |
                | --- | --- |
                | `invalid_request` | 400 |
                | `unauthorized` | 401 |
                | `invalid_token` | 401 |
                | `insufficient_scope` | 403 |
                | `forbidden` | 403 |
                | `not_found` | 404 |
                | `conflict` | 409 |
                | `unprocessable` | 422 |
                | `rate_limited` | 429 |
                | `misconfigured` | 500 |
                | `internal` | 500 |
            message:
              type: string
              description: >-
                A sentence written for a person. Safe to show, and safe to
                ignore — branch on `code`, not on this.
            request_id:
              type: string
              format: uuid
              description: The same value as the `X-Request-Id` header. Quote it when reporting a problem.
            details:
              type: object
              additionalProperties: true
              description: >-
                Present only when naming the offending thing is what makes the
                error useful. `{"editable": [...]}` on a patch with nothing
                patchable in it, `{"field": "default_currency"}` on a value of
                the wrong shape, `{"parameter": "include_archived"}` on an
                unreadable query flag, `{"header": "Idempotency-Key"}` on a key
                of the wrong length, or `{"scopes": [...]}` listing the
                catalogue when an unknown scope was named.
      examples:
        - error:
            code: insufficient_scope
            message: this token was not granted that
            request_id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
    OAuthError:
      type: object
      description: >-
        RFC 6749's error shape, used only by the `/oauth/*` endpoints. Note that
        `error` is a **string** here, not the object the rest of the API
        returns — this is the one place the service deliberately deviates from
        its own convention, because it is what every OAuth client library
        parses.
      required: [error, error_description]
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - server_error
        error_description:
          type: string
      examples:
        - error: invalid_grant
          error_description: that is not a code we issued

    # ── resources ──
...
    Settlement:
      type: object
      description: A payment one member says they made to another.
      required:
        - id
        - group_id
        - from_member_id
        - to_member_id
        - currency
        - amount
        - status
        - initiated_at
        - confirmed_at
      properties:
        id:
          type: string
          format: uuid
        group_id:
          type: string
          format: uuid
        from_member_id:
          type: string
          format: uuid
          description: Who paid.
        to_member_id:
          type: string
          format: uuid
          description: Who was paid. Only this person can confirm it.
        currency:
          $ref: '#/components/schemas/Currency'
        amount:
          $ref: '#/components/schemas/Money'
        status:
          $ref: '#/components/schemas/SettlementStatus'
        initiated_at:
          type: string
          format: date-time
          description: When the claim was recorded. This is the list's sort key.
        confirmed_at:
          type: [string, 'null']
          format: date-time
      examples:
        - id: 9f0e1d2c-3b4a-4958-8677-5a4b3c2d1e0f
          group_id: 8c1f2d6e-4a3b-4c5d-9e7f-1a2b3c4d5e6f
          from_member_id: 2c8e7fad-3b1c-4a9e-8d6f-7b5a4c3d2e10
          to_member_id: 1b9d6bcd-bbfd-4b2d-9b5d-ab8dfbbd4bed
          currency: INR
          amount: '120000'
          status: initiated
          initiated_at: '2026-09-16T08:03:44.900Z'
          confirmed_at: null
...
    SettlementPage:
      type: object
      required: [data, next_cursor]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Settlement'
        next_cursor:
          type: [string, 'null']
...
    CreateSettlementRequest:
      type: object
      required: [group_id, from_member_id, to_member_id, amount]
      properties:
        group_id:
          type: string
          format: uuid
        from_member_id:
          type: string
          format: uuid
          description: Who paid. Must differ from `to_member_id`.
        to_member_id:
          type: string
          format: uuid
          description: Who was paid.
        amount:
          allOf:
            - $ref: '#/components/schemas/UnsignedMoney'
          description: Strictly positive.
        currency:
          type: [string, 'null']
          description: Defaults to the group's currency when omitted.
        rail:
          type: string
          default: cash
          description: >-
            How the money moved. The finer rail, not a coarse enum — a rail this
            deployment has never heard of (Pix, PayNow, Wise) still records,
            because the stored category is derived from it.
          examples: ['upi']
        note:
          type: [string, 'null']
      examples:
        - group_id: 8c1f2d6e-4a3b-4c5d-9e7f-1a2b3c4d5e6f
          from_member_id: 2c8e7fad-3b1c-4a9e-8d6f-7b5a4c3d2e10
          to_member_id: 1b9d6bcd-bbfd-4b2d-9b5d-ab8dfbbd4bed
          amount: '120000'
          currency: INR
          rail: upi
          note: Settling up for dinner
...
    ApproveAuthorizationRequest:
      type: object
      description: >-
        The same parameters `GET /oauth/authorize` takes, as a JSON body, posted
        by the consent screen once the person has approved.
      required:
        [response_type, client_id, redirect_uri, scope, code_challenge, code_challenge_method]
      properties:
        response_type:
          type: string
          enum: [code]
        client_id:
          type: string
        redirect_uri:
          type: string
          format: uri
        scope:
          type: string
          description: Space-separated.
        code_challenge:
          type: string
          pattern: '^[A-Za-z0-9_-]{43}$'
        code_challenge_method:
          type: string
          enum: [S256]
        state:
          type: string
...